Rename lerp() parameters.
a, b, and t are not very descriptive. These have been renamed
to start, stop, and fraction, respectively.

I also corrected a few errors in docs that I saw.

Fixes: 142077662
Test: if it compiles, it should work.

+++ ui/ui-animation-core/src/main/java/androidx/animation/TransitionState.kt
@@ -91,7 +91,7 @@ class FloatPropKey : PropKey<Float> {
  */
 class IntPropKey : PropKey<Int> {
     override fun interpolate(a: Int, b: Int, fraction: Float) =
-        lerp(a, b, fraction).toInt()
+        lerp(a, b, fraction)
 }
 
 /**
